TestPersonKeyLong-Typ anpassen

+++ aero.minova.core.application.system.service/src/main/java/aero/minova/core/application/system/covid/test/print/domain/MeetingFormInformation.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class MeetingFormInformation {
 	private int testStreckKeyLong;
 	private LocalDate starttime;
-	private int testPersonKeyLong;
+	private long testPersonKeyLong;
 }
